# services/judge/api/openapi.yaml
openapi: 3.0.3
info:
  title: pokecalc judge API
  version: 0.1.0
  description: |
    判定(素早さ×ダメージ連動)の独立サービスの契約。
    「自分のポケモンが相手を素早さで抜けて、かつ使う技で倒せるか」を 1 回の request で判定する。
    judge は自分の契約を自分で持つ(ADR-0012)。ルートの api/openapi.yaml を $ref せず、
    同じ意味のスキーマをここで定義する。欄の意味の正はルートの api/openapi.yaml にある。
    契約の正は docs/adr/0700-judge-jd0-foundation.md(基盤)・
    docs/adr/0701-judge-jd1-outspeed-and-ko.md(判定)・
    docs/adr/0702-judge-jd2-field-effects.md(場の効果)・
    docs/adr/0703-judge-jd3-multiple-candidates.md(複数の相手候補)・
    docs/adr/0704-judge-jd4-counter-attack.md(相手の技を含めた返り討ち判定)と docs/judge-design.md。
servers:
  - url: /
paths:
  /healthz:
    get:
      operationId: health
      summary: Pod health check
      description: |
        Pod 自身が生きているかだけを返す。上流(pokedex-svc・calc-svc)の設定・疎通には依存しない
        (ADR-0700 §5): 上流が落ちている間に readiness を落とすと judge まで巻き込んで落ちるため。
        上流が使えないことは、判定の endpoint が 503 upstream_unavailable で表す。
      responses:
        '200':
          description: Ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
  /api/judge/healthz:
    get:
      operationId: publicHealth
      summary: Ingress smoke check
      description: |
        judge 自身の Ingress(/api/judge prefix。ADR-0700 §6-3)越しの疎通確認。
        本文は /healthz と同じで、上流には依存しない。
      responses:
        '200':
          description: Ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
  /api/judge/v1/outspeed-and-ko:
    post:
      operationId: outspeedAndKo
      summary: 素早さで抜けるか + 倒せるか + 返り討ちに遭わないか(相手候補ごとに)
      description: |
        自分(attacker)の調整と自分が使う技 1 つ、および相手候補(defenders。1〜6 件。**候補ごとに
        その候補が使う技 moveId を持つ**)を受け取り、候補ごとに「自分が相手より素早いか
        (outspeeds / speedTie)」「どちらが先に動くか(attackerMovesFirst / turnOrderTie)」
        「自分の技で相手を倒せるか(attackerKo)」「相手の技で自分が倒されるか(defenderKo)」を
        判定して matchups の配列で返す(ADR-0701 §1・ADR-0703 §1・ADR-0704 §1〜§3)。
        攻撃側と攻撃側の技は 1 つに固定する。

        先に動く側は **優先度(getMove の priority)が違えば優先度が高い方**で、素早さもトリックルームも
        見ない。優先度が同じときだけ素早さ(outspeeds。トリックルーム反映済み)で決まり、
        優先度も素早さも同じなら turnOrderTie が true になってどちらが先かは決まらない(ADR-0704 §2)。
        judge は「勝てる / 負ける」の真偽値に丸めず、双方の確定数と行動順をそのまま返す。

        matchups は defenders と**同じ順序・同じ件数**で、各行は defenderIndex(defenders 配列の位置。
        0 始まり)を持つ。attackerSpeed はどの行でも同じ値になるが、行を自己完結させるために
        各行が持つ(ADR-0703 §2)。同じ speciesKey が候補に重複していても取りまとめず、
        候補の数だけ素直に上流を引く(ADR-0703 §6。技も同じで、同じ moveId をまとめない)。

        素早さは 実数値(engine.RealStats)→ ランク補正 → 素早さ補正(追い風 ×2・こだわりスカーフ ×1.5)の
        順で求める(ADR-0701 §2・ADR-0702 §2)。補正は 4096 基準で 1 つに連結してから 1 回だけ
        五捨五超入する(各補正ごとに丸めない)。状態異常(麻痺など)による素早さの変化は扱わないため、
        Individual に status は無い。ko は calc-svc の計算結果をそのまま転記する(judge は確定数を再計算しない)。

        場の効果は 2 つの欄に分かれる(ADR-0702 §1)。ダメージに効く weather / terrain / screens は
        field に入れ、judge は解釈せず calc-svc へそのまま転送する。素早さにしか効かない
        トリックルーム・追い風は speedField に入れ、judge だけが解釈して calc-svc には送らない。
        field の attackerScreens / defenderScreens は「自分の側 / 相手の側」に張られた壁を表すので、
        **逆方向(相手 → 自分)のダメージ計算のときだけ 2 つを入れ替えて** calc-svc に送る
        (壁は場の各側にあり、どちらが殴るかで場所は変わらない。weather / terrain は入れ替えない。ADR-0704 §4)。

        検査順(ADR-0701 §5 を ADR-0703 §4・ADR-0704 §5 が拡張した形。上流は逐次で呼ぶ):
        ヘッダー (400) → body が 1 つの JSON か・上限 8 KiB (400 / 413)
        → defenders の件数が 1〜6 か (400)
        → attacker と各 defender の sp・ranks・format・必須文字列(候補の moveId を含む)の範囲と、
           ID の形式(moveId / natureId が MoveId / NatureId の pattern・maxLength に合うか。ADR-0706 §1)
           (400。attacker を先に見て、次に defenders を index 昇順。最初に範囲外だった候補の index で止める。
            ここまで上流を 1 回も呼ばない。形式が合わない ID は 503 upstream_unavailable にならない。ADR-0706 §3)
        → 性格の一覧 (1 リクエストにつき 1 回だけ。503)
        → natureId が一覧に無い (422 unknown_nature。attacker → defenders を index 昇順に見て最初の 1 件)
        → attacker の種族 (422 unknown_species / 503)
        → attacker の技 (422 unknown_move / 503)
        → 各 defender の【種族 → 技】(index 昇順。最初に失敗した候補で打ち切る)
        → 各 defender の【順方向のダメージ計算 → 逆方向のダメージ計算】
           (index 昇順。最初に失敗した候補で打ち切る。400 / 503) → 200。

        1 リクエストの上流呼び出しは 性格 1 + 種族 (1+N) + 技 (1+N) + ダメージ計算 2N の
        3+4N 回になる(N は候補数)。**攻撃側の技も上流で解決するようになったため、未知の技は
        calc-svc に届く前に 422 unknown_move になる**(JD3 までは calc-svc の 400 が
        invalid_request に畳まれていた。ADR-0704 §7)。

        **エラーは最初に失敗した候補で全体を打ち切る**(部分的な成功は返さない。ADR-0703 §3)。
        どの候補で失敗したかは、エラーの message に defenders[<index>] の形で入る
        (上流の URL・本文は入らない。ADR-0700 §3 は保たれる)。
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutspeedAndKoRequest'
      responses:
        '200':
          description: 相手候補ごとの素早さ・行動順の比較結果と双方向の確定数(defenders と同じ順序・同じ件数)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutspeedAndKoResponse'
        '400':
          description: |
            X-Device-Id / X-Session-Id が無い・空、body が 1 つの JSON オブジェクトでない・未知の欄がある、
            必須の欄が無い(候補の moveId を含む)、defenders が 1 件未満・6 件超、sp / ranks / format が範囲外、
            または calc-svc が計算要求を受け付けなかった(invalid_request。ADR-0701 §6・ADR-0703 §4)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: body が 8 KiB を超えている(request_too_large)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: |
            speciesKey が pokedex-svc のマスタに無い(unknown_species)、
            moveId が技のマスタに無い(unknown_move。攻撃側・候補側のどちらも。ADR-0704 §6)、
            または natureId が性格の一覧に無い(unknown_nature)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: |
            pokedex-svc / calc-svc が未設定・接続できない・タイムアウト・5xx・契約に合わない応答
            (upstream_unavailable)。/healthz はこのとき 200 のまま(ADR-0700 §5)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 想定外の内部エラー(internal_error)。message は固定文言で、内部の詳細を返さない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  parameters:
    DeviceId:
      name: X-Device-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 1
    SessionId:
      name: X-Session-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 1
  schemas:
    Health:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok
    SpeciesKey:
      description: '{図鑑番号4桁}-{フォルム3桁}。意味の正はルートの api/openapi.yaml の SpeciesKey。'
      type: string
      pattern: '^[0-9]{4}-[0-9]{3}$'
      example: '0445-000'
    MoveId:
      description: |
        技 ID(Showdown ID: 小文字英数をハイフンで区切る)。意味の正はルートの api/openapi.yaml の Move.id。
        judge はこの ID を GET /api/pokedex/moves/{key} の path 要素に埋めるので、形式に合わない値は
        上流を呼ぶ前に 400 invalid_request で断る(ADR-0706 §1・§3)。形式は合うがマスタに無いときは
        422 unknown_move(ADR-0704 §6)。
      type: string
      minLength: 1
      maxLength: 64
      pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
      example: flamethrower
    NatureId:
      description: |
        性格 ID(Showdown ID: 小文字英数をハイフンで区切る)。GET /api/pokedex/natures の一覧で
        補正する能力に解決する(ADR-0701 §4)。URL には埋めないが、同じ由来の ID を欄ごとに違う
        厳しさで通さないため MoveId と同じ形式で検査する(ADR-0706 §5)。形式に合わない値は
        上流を呼ぶ前に 400 invalid_request、一覧に無いときは 422 unknown_nature。
      type: string
      minLength: 1
      maxLength: 64
      pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
      example: jolly
    Format:
      description: 対戦形式。calc-svc にそのまま渡す。
      type: string
      enum:
        - single
        - double
      default: single
    StatBlock:
      description: |
        6 ステータスの値。judge では Individual.sp(能力ポイント。各 0..32・合計 <= 66)に使う。
        意味の正はルートの api/openapi.yaml の StatBlock。
      type: object
      required:
        - hp
        - atk
        - def
        - spa
        - spd
        - spe
      properties:
        hp:
          type: integer
          minimum: 0
          maximum: 32
        atk:
          type: integer
          minimum: 0
          maximum: 32
        def:
          type: integer
          minimum: 0
          maximum: 32
        spa:
          type: integer
          minimum: 0
          maximum: 32
        spd:
          type: integer
          minimum: 0
          maximum: 32
        spe:
          type: integer
          minimum: 0
          maximum: 32
    RankBlock:
      description: |
        ランク補正(-6..+6)。HP は持たない。judge は「技の追加効果を適用した後のランク」を
        呼び出し側が入れたものとして受け取る(ADR-0700 §6-5)。
      type: object
      properties:
        atk:
          type: integer
          minimum: -6
          maximum: 6
          default: 0
        def:
          type: integer
          minimum: -6
          maximum: 6
          default: 0
        spa:
          type: integer
          minimum: -6
          maximum: 6
          default: 0
        spd:
          type: integer
          minimum: -6
          maximum: 6
          default: 0
        spe:
          type: integer
          minimum: -6
          maximum: 6
          default: 0
    Individual:
      description: |
        判定に使う個体。欄は docs/judge-design.md §3 JD1 の列挙そのまま。
        status(状態異常)と teraType は JD1 では受け取らない(ADR-0701 §2)。
      type: object
      required:
        - speciesKey
        - natureId
        - sp
      properties:
        speciesKey:
          $ref: '#/components/schemas/SpeciesKey'
        natureId:
          $ref: '#/components/schemas/NatureId'
        sp:
          $ref: '#/components/schemas/StatBlock'
        ranks:
          $ref: '#/components/schemas/RankBlock'
        abilityId:
          description: 特性 ID。judge は解釈せず calc-svc にそのまま渡す。
          type: string
          nullable: true
        itemId:
          description: |
            持ち物 ID。judge は calc-svc にそのまま渡すほか、こだわりスカーフの ID
            (既定 choicescarf。環境変数 JUDGE_CHOICE_SCARF_ITEM_ID で上書きできる。ADR-0701 §3)
            と一致するときだけ素早さに ×1.5 を掛ける。
          type: string
          nullable: true
    DefenderCandidate:
      description: |
        相手候補 1 件(ADR-0704 §1)。Individual の全欄に加えて、**この候補が撃ち返してくる技
        moveId を必ず持つ**。judge はその技の優先度(GET /api/pokedex/moves/{key} の priority)で
        先に動く側を決め、その技で自分がどれだけ削られるか(defenderKo)を計算する。

        Individual を allOf で継承せずに欄を書き下しているのは意図的で、生成される型を素の
        オブジェクトのまま保つため(ADR-0704 §1)。欄の意味は Individual と同じで、正は
        ルートの api/openapi.yaml にある。attacker は技を 1 つだけ持ち(request 直下の moveId)、
        候補は候補ごとに違う技を持つので、型を分けて「候補は必ず技を持つ」を required で表す。
      type: object
      required:
        - speciesKey
        - natureId
        - sp
        - moveId
      properties:
        speciesKey:
          $ref: '#/components/schemas/SpeciesKey'
        natureId:
          $ref: '#/components/schemas/NatureId'
        sp:
          $ref: '#/components/schemas/StatBlock'
        ranks:
          $ref: '#/components/schemas/RankBlock'
        abilityId:
          description: 特性 ID。judge は解釈せず calc-svc にそのまま渡す。
          type: string
          nullable: true
        itemId:
          description: |
            持ち物 ID。judge は calc-svc にそのまま渡すほか、こだわりスカーフの ID
            (既定 choicescarf。ADR-0701 §3)と一致するときだけ素早さに ×1.5 を掛ける。
          type: string
          nullable: true
        moveId:
          description: |
            この候補が使う技(1 つ)。優先度は GET /api/pokedex/moves/{key} で引き、
            この技によるダメージは calc-svc を逆方向(この候補が攻撃側・自分が防御側)で
            呼んで求める(ADR-0704 §4)。マスタに無ければ 422 unknown_move。
            形式が MoveId に合わない値は上流を呼ぶ前に 400 invalid_request で、
            message は defenders[<index>] を示す(ADR-0706 §3・§4)。
          allOf:
            - $ref: '#/components/schemas/MoveId'
    Weather:
      type: string
      enum:
        - none
        - sun
        - rain
        - sand
        - snow
      default: none
    Terrain:
      type: string
      enum:
        - none
        - electric
        - grassy
        - misty
        - psychic
      default: none
    Screens:
      type: object
      properties:
        reflect:
          type: boolean
          default: false
        lightScreen:
          type: boolean
          default: false
        auroraVeil:
          type: boolean
          default: false
    FieldState:
      description: |
        ダメージに効く場の状態。judge は解釈せず calc-svc の field にそのまま転送する(省略時は送らない)。
        意味の正はルートの api/openapi.yaml の FieldState。
        素早さにしか効かないトリックルーム・追い風はここに入れない(speedField。ADR-0702 §1)。
        すべての相手候補に同じ field が使われる(候補ごとに天候を変えることはできない。ADR-0703 §5)。
      type: object
      properties:
        weather:
          $ref: '#/components/schemas/Weather'
        terrain:
          $ref: '#/components/schemas/Terrain'
        attackerScreens:
          $ref: '#/components/schemas/Screens'
        defenderScreens:
          $ref: '#/components/schemas/Screens'
    SpeedField:
      description: |
        素早さの判定にだけ効く場の効果(ADR-0702 §1)。judge が自分で解釈し、calc-svc には送らない。
        トリックルーム・追い風はダメージに関与せず、calc-svc は場の効果として weather / terrain /
        screens しか理解しないため、calc-svc へ転送する field とは別の欄にする。
        speedField 自体を省略した場合も、個々の欄を省略した場合も、すべて false(場の効果なし)
        として扱う。したがって speedField を送らない request の挙動は JD1 と同じになる。
        **1 リクエストに 1 つだけで、すべての相手候補に同じように適用される**(ADR-0703 §5)。
        候補ごとに追い風の有無を変えることはできない。
      type: object
      properties:
        trickRoom:
          description: |
            トリックルームがかかっているか。トリックルームは素早さの実数値そのものを変えないため、
            attackerSpeed / defenderSpeed は変わらない。変わるのは outspeeds(どちらが先に動くか)の
            比較の向きだけで、true のときは「遅い方が先に動く」= attackerSpeed < defenderSpeed に
            なる(ADR-0702 §3)。speedTie(同速)は反転しない。
          type: boolean
          default: false
        attackerTailwind:
          description: |
            自分の側に追い風がかかっているか。追い風は実数値そのものを ×2 する(ADR-0702 §2)ため、
            attackerSpeed にも反映される。
          type: boolean
          default: false
        defenderTailwind:
          description: |
            相手の側に追い風がかかっているか。相手の戦闘中の素早さを ×2 する。
            すべての相手候補に同じように適用される(ADR-0703 §5)。
          type: boolean
          default: false
    OutspeedAndKoRequest:
      type: object
      required:
        - format
        - attacker
        - defenders
        - moveId
      properties:
        format:
          $ref: '#/components/schemas/Format'
        attacker:
          description: 自分の個体。攻撃側は 1 つに固定する(ADR-0700 §6-2・ADR-0703 §1)。
          allOf:
            - $ref: '#/components/schemas/Individual'
        defenders:
          description: |
            相手候補の一覧(ADR-0703 §1)。1〜6 件。判定は候補ごとに行われ、
            matchups がこの配列と同じ順序・同じ件数で返る。要素は DefenderCandidate で、
            **候補ごとに自分の技 moveId を持つ**(JD4 の破壊的変更。ADR-0704 §1・§8)。
            上限 6 件は「手持ちの数」に合わせた慣習(services/balance の members / threats と同じ)で、
            1 リクエストの逐次の上流呼び出し数(性格 1 + 種族 (1+N) + 技 (1+N) + 計算 2N)を
            読める範囲に抑えるために置く。同じ speciesKey が重複していても取りまとめない。
          type: array
          minItems: 1
          maxItems: 6
          items:
            $ref: '#/components/schemas/DefenderCandidate'
        moveId:
          description: |
            自分(attacker)が使う技(1 つ)。すべての候補に対して同じ技で判定する。
            JD4 からは優先度を引くために GET /api/pokedex/moves/{key} でも解決するので、
            マスタに無ければ calc-svc に届く前に 422 unknown_move になる(ADR-0704 §7)。
            形式が MoveId に合わない値は上流を呼ぶ前に 400 invalid_request で、
            message は attacker を示す(候補の index を騙らない。ADR-0706 §4)。
          allOf:
            - $ref: '#/components/schemas/MoveId'
        field:
          $ref: '#/components/schemas/FieldState'
        speedField:
          $ref: '#/components/schemas/SpeedField'
    KOChance:
      description: |
        確定数 / 乱数 n 発。calc-svc の KOChance をそのまま転記する(judge は再計算しない)。
        意味の正はルートの api/openapi.yaml の KOChance と ADR-0006・ADR-0010。
        engine の生値 chancePercent は画面に出す値ではないので judge は返さない。
      type: object
      required:
        - hits
        - guaranteed
        - displayChancePercent
      properties:
        hits:
          description: 最大ダメージで倒すのに必要な攻撃回数(0 = 倒せない)。
          type: integer
        guaranteed:
          description: 最小ダメージでも hits 回で倒せるなら true(確定 n 発)。
          type: boolean
        displayChancePercent:
          description: 画面に出す「hits 回で倒せる確率(%)」。小数第1位(ADR-0010 §3)。
          type: number
          format: double
    UnsupportedMark:
      description: |
        「この確定数は正しくない可能性がある」印 1 つ(ADR-0123・ADR-0708)。engine が正しく計算できない
        技の機構・持ち物・特性に、数値は通常の式のまま付く(400 で拒否されない)。
        意味・条件・並びの正はルートの api/openapi.yaml の UnsupportedMark と ADR-0123 で、
        judge は calc-svc が返した値を**そのまま・同じ順で**中継するだけである(解釈・並べ替え・
        重複除去・真偽値への丸めをしない。ADR-0708 §4)。judge が自分の契約に同じ定義を持つのは、
        ルートの契約を $ref せず契約を独立に版管理するため(ADR-0012・ADR-0706 §2 の前例)。
      type: object
      required:
        - target
        - reason
        - id
      properties:
        target:
          description: |
            印の対象。attacker / defender は**その計算から見た**役割で、judge の自分・相手とは
            一致しないことがある(ADR-0708 §5)。attackerKoUnsupported(順方向)では
            attacker_* = 自分・defender_* = その候補、defenderKoUnsupported(逆方向)では
            attacker_* = その候補・defender_* = 自分を指す。
          type: string
          enum:
            - move
            - attacker_item
            - attacker_ability
            - defender_item
            - defender_ability
        reason:
          description: |
            印の理由。技は機構の値(13 種)か zero_power(威力 0 の攻撃技)、持ち物・特性は
            unsupported_effect(効果スキーマで表せない)。judge はこの値を検査せず、
            この列挙に無い値もそのまま中継する(engine が理由を足したときに judge の版で落とさない。
            ADR-0708 §4・§6。契約は説明で、judge は印の意味を持たない)。
          type: string
          enum:
            - alt_defense_stat
            - alt_offense_stat
            - always_crit
            - effectiveness_change
            - field_specific
            - fixed_damage
            - ignore_defense_ranks
            - move_specific
            - multi_hit
            - ohko
            - priority_change
            - type_change
            - variable_power
            - zero_power
            - unsupported_effect
        id:
          description: 印が付いた技・持ち物・特性の ID(calc-svc が返したまま)。
          type: string
    OutspeedAndKoResponse:
      type: object
      required:
        - matchups
      properties:
        matchups:
          description: |
            相手候補ごとの判定結果(ADR-0703 §2)。request の defenders と**同じ順序・同じ件数**で、
            i 番目の行の defenderIndex は i になる。部分的な成功は返さない(どれか 1 つの候補で
            失敗したら request 全体がエラーになる。ADR-0703 §3)。
          type: array
          minItems: 1
          maxItems: 6
          items:
            $ref: '#/components/schemas/Matchup'
    Matchup:
      description: |
        相手候補 1 件ぶんの判定結果(ADR-0703 §2)。attackerSpeed はどの行でも同じ値になるが、
        行だけを見て「何対何で抜けているか」が分かるように各行が持つ
        (ルートの api/openapi.yaml の BulkCalcRow が各行に防御側の実数値を持たせているのと同じ形)。
      type: object
      required:
        - defenderIndex
        - outspeeds
        - speedTie
        - attackerSpeed
        - defenderSpeed
        - attackerMovePriority
        - defenderMovePriority
        - attackerMovesFirst
        - turnOrderTie
        - attackerKo
        - defenderKo
        - attackerKoUnsupported
        - defenderKoUnsupported
      properties:
        defenderIndex:
          description: |
            この行が対応する request の defenders の位置(0 始まり)。matchups は defenders と
            同じ順序で返るので i 番目は必ず i になるが、行を並べ替えて表示する画面が
            対応を取り違えないように明示する。
          type: integer
          minimum: 0
          maximum: 5
        outspeeds:
          description: |
            **素早さの比較で**自分が先に動く側か(ADR-0700 §6-1・ADR-0702 §3。値の意味は
            JD1〜JD3 から変わらない)。トリックルームが無ければ attackerSpeed > defenderSpeed、
            speedField.trickRoom が true なら attackerSpeed < defenderSpeed。同速は false で、
            speedTie が true になる(真偽値 1 つに丸めない)。
            技の優先度まで含めた最終的な行動順は attackerMovesFirst / turnOrderTie で、
            優先度が違えばこの欄と食い違う(ADR-0704 §2)。
          type: boolean
        speedTie:
          description: |
            双方の戦闘中の素早さが等しいか。outspeeds と同時に true にはならない。
            同速はトリックルームの有無に関わらず行動順が決まらないため、trickRoom では反転しない。
          type: boolean
        attackerSpeed:
          description: |
            自分の戦闘中の素早さ(ランク・追い風・こだわりスカーフ適用後)。
            トリックルームは実数値を変えないので、この値には現れない(ADR-0702 §2)。
            攻撃側は 1 つに固定なので、すべての matchups で同じ値になる。
          type: integer
        defenderSpeed:
          description: この候補の戦闘中の素早さ(ランク・追い風・こだわりスカーフ適用後)。
          type: integer
        attackerMovePriority:
          description: |
            自分が使う技(request 直下の moveId)の優先度。GET /api/pokedex/moves/{key} の
            priority をそのまま転記する(ADR-0704 §2・§3)。画面が「なぜ先に動くのか
            (速いからか、先制技だからか)」を説明できるように返す。
          type: integer
        defenderMovePriority:
          description: この候補が使う技(defenders[i].moveId)の優先度。同じく priority の転記。
          type: integer
        attackerMovesFirst:
          description: |
            自分が先に動くか(ADR-0704 §2)。**優先度が違えば優先度が高い方が先**で、
            素早さもトリックルームも見ない。優先度が同じときだけ outspeeds(トリックルーム
            反映済み)に従う。turnOrderTie が true のときは false になる
            (どちらが先か決まらないので「自分が先」とは言えない)。
          type: boolean
        turnOrderTie:
          description: |
            優先度も戦闘中の素早さも同じで、どちらが先に動くかが決まらないか。
            attackerMovesFirst と同時に true にはならない(真偽値 1 つに丸めない。ADR-0700 §6-1)。
            優先度が違えば素早さが同じでも false になる(速さの同速 speedTie とは別の欄)。
          type: boolean
        attackerKo:
          description: |
            自分の技がこの候補に与えるダメージの確定数(JD3 までの ko。ADR-0704 §3 で改名)。
            calc-svc を順方向(自分が攻撃側)で呼んだ結果をそのまま転記する。
          allOf:
            - $ref: '#/components/schemas/KOChance'
        defenderKo:
          description: |
            この候補の技(defenders[i].moveId)が自分に与えるダメージの確定数(ADR-0704 §3)。
            calc-svc を逆方向(この候補が攻撃側・自分が防御側)で呼んだ結果をそのまま転記する。
            逆方向では field の attackerScreens / defenderScreens を入れ替えて送る(ADR-0704 §4)。
            行動順に関わらず必ず計算する(自分が先に動いて倒しきれなかったときの被害も知りたいため)。
          allOf:
            - $ref: '#/components/schemas/KOChance'
        attackerKoUnsupported:
          description: |
            attackerKo(順方向の計算。自分の技 → この候補)に付いた「正しく計算できていない可能性がある」印
            (ADR-0708 §1)。calc-svc の CalcResult.unsupported をそのまま・同じ順で中継する。
            **印が無いときは空配列**(null にも欄の欠落にもしない。ADR-0708 §3)。
            画面はこれが空でないとき、attackerKo を「確定した数」として見せない
            (「この確定数は当てにならないかもしれない」旨を添える。文言は画面の持ち物)。
            自分の技が多段技・威力変動・固定ダメージのとき、またはこの計算で効く持ち物・特性が
            engine の効果スキーマで表せないときに入る。
            target の attacker_* は自分・defender_* はこの候補を指す(ADR-0708 §5)。
          type: array
          items:
            $ref: '#/components/schemas/UnsupportedMark'
        defenderKoUnsupported:
          description: |
            defenderKo(逆方向の計算。この候補の技 → 自分)に付いた印(ADR-0708 §1)。
            attackerKoUnsupported と同じ形で、**印が無いときは空配列**。
            順方向の印と**まとめない**(どちらの確定数が疑わしいかを画面が区別できるように、
            方向ごとに分けたまま返す。ADR-0708 §4)。
            逆方向では役割が入れ替わるので、target の attacker_* は**この候補**・
            defender_* は**自分**を指す(ADR-0708 §5。自分が持つ防御側で効く持ち物の印はこちらに入る)。
          type: array
          items:
            $ref: '#/components/schemas/UnsupportedMark'
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
    ErrorCode:
      description: |
        エラーの区分。judge は上流の事情(HTTP のステータス・接続エラーの文面・URL)をそのまま返さず、
        ADR-0700 §3・ADR-0701 §6 の対応表でこの列挙に畳む。
        invalid_request: ヘッダー・request body が契約に合わない、または calc-svc が計算要求を受け付けなかった。
        unknown_species: speciesKey が pokedex-svc のマスタに無い。
        unknown_move: moveId が pokedex-svc の技のマスタに無い(攻撃側・候補側のどちらも。ADR-0704 §6)。
        unknown_nature: natureId が性格の一覧に無い。
        request_too_large: request body が上限(8 KiB)を超えている。
        upstream_unavailable: pokedex-svc / calc-svc が未設定・接続できない・タイムアウト・5xx・契約に合わない応答。
        internal_error: 想定外の内部エラー(message は固定文言で、内部の詳細を返さない)。
      type: string
      enum:
        - invalid_request
        - unknown_species
        - unknown_move
        - unknown_nature
        - request_too_large
        - upstream_unavailable
        - internal_error